+++ Server/AutomationActions/Gestures/EnterText.m
@@ -3,7 +3,8 @@
 #import "Testmanagerd.h"
 #import "ThreadUtils.h"
 #import "EnterText.h"
-#import "Application.h"
+#import "CBXDevice.h"
+#import "TextInputFirstResponderProvider.h"
 
 @implementation EnterText
 
@@ -17,6 +18,7 @@
     return @[];
 }
 
+// Callers are responsible for checking that a keyboard is visible.
 + (Gesture *)executeWithGestureConfiguration:(GestureConfiguration *)gestureConfig
                                        query:(Query *)query
                                   completion:(CompletionBlock)completion {
@@ -26,28 +28,38 @@
 
     NSString *string = gestureConfig[CBX_STRING_KEY];
 
-// Original implementation - not working on iOS 10 physical devices
-// https://xamarin.atlassian.net/browse/TCFW-333
-//
-//    [ThreadUtils runSync:^(BOOL *setToTrueWhenDone, NSError *__autoreleasing *err) {
-//        [[Testmanagerd get] _XCT_sendString:string
-//         maximumFrequency:CBX_DEFAULT_SEND_STRING_FREQUENCY
-//                                 completion:^(NSError *e) {
-//            *err = e;
-//            *setToTrueWhenDone = YES;
-//        }];
-//    } completion:completion];
-
-    // This is working on iOS 9 - 10 sims and physical devices.
-    XCUIApplication *application = [Application currentApplication];
-    if ([application lastSnapshot] == nil) {
-        [[application applicationQuery] elementBoundByIndex:0];
-        [application resolve];
+    __block NSError *outerError = nil;
+    if ([[CBXDevice sharedDevice] isArm64]) {
+        NSLog(@"Device has arm64 architecture: %@", [[CBXDevice sharedDevice] armVersion]);
+        [ThreadUtils runSync:^(BOOL *setToTrueWhenDone) {
+            [[Testmanagerd get]
+             _XCT_sendString:string
+             maximumFrequency:CBX_DEFAULT_SEND_STRING_FREQUENCY
+             completion:^(NSError *innerError) {
+                 outerError = innerError;
+                 *setToTrueWhenDone = YES;
+                 NSString *message;
+                 message = [NSString stringWithFormat:@"Encountered an error typing text: %@",
+                            innerError];
+                 if (innerError) {
+                     @throw [CBXException withMessage:message];
+                 }
+             }];
+        }];
+    } else {
+        NSLog(@"Device does not have arm64 architecture: %@",
+              [[CBXDevice sharedDevice] armVersion]);
+        // Gestures need to be automatic - a gesture must complete before the next
+        // gesture started.  Depending on the internal implementation of typeText:,
+        // this should block other gestures.
+        @synchronized([Testmanagerd get]) {
+            TextInputFirstResponderProvider *provider = [TextInputFirstResponderProvider new];
+            XCUIElement *firstResponder = [provider firstResponderOrApplication];
+            [firstResponder typeText:string];
+        }
     }
 
-    [application typeText:string];
-
-    completion(nil);
+    completion(outerError);
     return nil;
 }