+++ Server/NSXPCConnection.h
@@ -55,10 +55,12 @@
 @class XCSynthesizedEventRecord;
 @class XCElementSnapshot;
 
+/*
 @protocol XCTestManager_TestsInterface
 - (void)_XCT_receivedAccessibilityNotification:(int)arg1 withPayload:(NSData *_Nonnull)arg2;
 - (void)_XCT_applicationWithBundleID:(NSString *_Nonnull)arg1 didUpdatePID:(int)arg2 andState:(unsigned long long)arg3;
 @end
+*/
 
 //TODO Just add the header
 @interface XCAccessibilityElement : NSObject