+++ TestApp/App/Misc/TextInputController.h
@@ -1,0 +1,7 @@
+
+#import <UIKit/UIKit.h>
+
+@interface TextInputController : UIViewController
+<UITextFieldDelegate, UITextViewDelegate>
+
+@end