+++ TestApp/App/MiscViewController.h
@@ -1,6 +1,0 @@
-
-#import <UIKit/UIKit.h>
-
-@interface MiscViewController : UIViewController<UITextFieldDelegate>
-
-@end