+++ bin/ci/jenkins.sh
@@ -1,15 +1,23 @@
 #!/usr/bin/env bash
 
+
+export DEVELOPER_DIR=/Xcode/8.1/Xcode.app/Contents/Developer
+
+xcrun simctl help >/dev/null 2>&1
+xcrun simctl help >/dev/null 2>&1
+xcrun simctl help >/dev/null 2>&1
+
 set -e
 
 bundle install
 make docs
+bin/ci/install-keychain.sh
 make app-agent
 make unit-tests
-bin/ci/make-ipa-agent.sh
-bin/ci/make-test-ipa.sh
+make ipa-agent
+make test-ipa
 
-# app-unit is required for cucumber tests.
-#make test-app
-#bundle exec bin/ci/cucumber.rb
+# TestApp is required for Cucumber tests.
+make test-app
+bundle exec bin/ci/cucumber.rb